An initial commit defining some master-related header files, data structures and function.
slaves.h introduces two arrays - one containing addresses of all connected slaves (it would be iterated over in mss_run_master during realisation of round-robin algorithm) and packet_count in which last obtained packet number is contained. packet_count size is redundant - it has size of MAX_ADDRESS+1, which means most of the array is unused. It is a speed for structure size tradeoff, as such construction allows instant access to the counter of a given slave.

mss_stop_master() was introduced in order to allow stoppage of mss-master when program runs in multithreaded program.