Separate DAT counting for each pair of slaves.
Now each pair of slaves has individual packet counter. Therefore each slave holds an array of counters for outcoming packets, indexed by packet destination. It is convinient, as it allows slaves not to track traffic on the bus (i.e. go offline) and care only about messages that are addressed to him.

Plus SDN packets are no longer counted. And the are not counted, because they need not to be counted.